run TestCache on Linux
