+++ sdk/azidentity/cache/cache_test.go
@@ -1,6 +1,6 @@
-//go:build go1.18 && (darwin || linux || windows)
+//go:build go1.18 && (linux || windows)
 // +build go1.18
-// +build darwin linux windows
+// +build linux windows
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
@@ -8,68 +8,63 @@
 package cache
 
 import (
-	"errors"
+	"context"
 	"path/filepath"
-	"sync"
+	"strings"
 	"testing"
 
-	"github.com/AzureAD/microsoft-authentication-extensions-for-go/cache/accessor"
-	"github.com/AzureAD/microsoft-authentication-extensions-for-go/cache/accessor/file"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/stretchr/testify/require"
 )
 
-func TestNew(t *testing.T) {
-	errBefore := storageError
-	onceBefore := once
-	storageBefore := storage
-	tryStorageBefore := tryStorage
-	t.Cleanup(func() {
-		once = onceBefore
-		storage = storageBefore
-		storageError = errBefore
-		tryStorage = tryStorageBefore
-	})
-	for _, expectedErr := range []error{nil, errors.New("it didn't work")} {
-		name := "storage error"
-		if expectedErr == nil {
-			name = "no storage error"
-		}
-		t.Run(name, func(t *testing.T) {
-			once = &sync.Once{}
-			storage = func(string) (accessor.Accessor, error) {
-				p := filepath.Join(t.TempDir(), t.Name())
-				return file.New(p)
-			}
-			storageError = nil
-			tries := 0
-			tryStorage = func() {
-				tries++
-				storageError = expectedErr
-			}
-			wg := &sync.WaitGroup{}
-			ch := make(chan error, 1)
-			for i := 0; i < 50; i++ {
-				wg.Add(1)
-				go func() {
-					defer wg.Done()
-					if _, err := New(&Options{Name: t.Name()}); err != nil {
-						select {
-						case ch <- err:
-						default:
-						}
-					}
-				}()
-			}
-			wg.Wait()
-			select {
-			case err := <-ch:
-				if expectedErr == nil {
-					t.Fatal(err)
+var ctx = context.Background()
+
+func TestCache(t *testing.T) {
+	before := cacheDir
+	t.Cleanup(func() { cacheDir = before })
+	cacheDir = func() (string, error) { return t.TempDir(), nil }
+	for _, test := range []struct {
+		credential func(azidentity.Cache) (azcore.TokenCredential, error)
+		name       string
+	}{
+		{
+			func(c azidentity.Cache) (azcore.TokenCredential, error) {
+				opts := azidentity.ClientSecretCredentialOptions{
+					Cache:         c,
+					ClientOptions: policy.ClientOptions{Transport: &mockSTS{}},
 				}
-				require.EqualError(t, err, expectedErr.Error())
-			default:
-			}
-			require.Equal(t, 1, tries, "tryStorage was called more than once")
+				return azidentity.NewClientSecretCredential("tenantID", "clientID", "secret", &opts)
+			},
+			"confidential",
+		},
+		{
+			func(c azidentity.Cache) (azcore.TokenCredential, error) {
+				opts := azidentity.DeviceCodeCredentialOptions{
+					Cache:         c,
+					ClientOptions: policy.ClientOptions{Transport: &mockSTS{}},
+				}
+				return azidentity.NewDeviceCodeCredential(&opts)
+			},
+			"public",
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			name := strings.ReplaceAll(t.Name(), string(filepath.Separator), "_")
+			cache, err := New(&Options{Name: name})
+			require.NoError(t, err)
+			cred, err := test.credential(cache)
+			require.NoError(t, err)
+			tro := policy.TokenRequestOptions{Scopes: []string{"scope"}}
+			tk, err := cred.GetToken(ctx, tro)
+			require.NoError(t, err)
+
+			cred2, err := test.credential(cache)
+			require.NoError(t, err)
+			tk2, err := cred2.GetToken(ctx, tro)
+			require.NoError(t, err)
+			require.Equal(t, tk.Token, tk2.Token)
 		})
 	}
 }
